Implemented all services pending to test

+++ src/main/java/me/team/usercrud/application/user/impl/UserDeleteServiceImpl.java
@@ -2,6 +2,7 @@ package me.team.usercrud.application.user.impl;
 
 import lombok.AllArgsConstructor;
 import me.team.usercrud.application.user.UserDeleteService;
+import me.team.usercrud.application.user.exceptions.UserNotFoundException;
 import me.team.usercrud.domain.User;
 import me.team.usercrud.domain.UserRepository;
 import org.springframework.stereotype.Service;
@@ -16,11 +17,20 @@ public class UserDeleteServiceImpl implements UserDeleteService {
 
     @Override
     public void deleteUserById(UUID userID) {
+        if (!userRepository.existsById(userID)) {
+            throw new UserNotFoundException(userID);
+        }
+
         userRepository.deleteById(userID);
     }
 
     @Override
     public void deleteUser(User user) {
-        userRepository.deleteById(user.getId());
+        final var userId = user.getId();
+        if (!userRepository.existsById(userId)) {
+            throw new UserNotFoundException(userId);
+        }
+
+        userRepository.deleteById(userId);
     }
 }
